requirements.txt
amqp==1.4.7
anyjson==0.3.3
billiard==3.3.0.21
celery==3.1.19
django-celery==3.1.17
django-filter==0.11.0
djangorestframework==3.3.1
kombu==3.0.29
Markdown==2.6.5
pytz==2015.7
redis==2.10.5
wheel==0.24.0
